+++ include/matrix/metrics.hpp
@@ -2,18 +2,14 @@
 
 #include "products.hpp"
 
-namespace math
-{
+namespace math {
 
-template <typename T>
-double norm(const DynamicVector<T>& vector)
-{
-    return sqrt(dot(vector,vector));
+template <typename T> double norm(const DynamicVector<T> &vector) {
+  return sqrt(dot(vector, vector));
 }
 
-template <typename T, int N>
-double norm(const StaticVector<T,N>& vector)
-{
-    return sqrt(static_cast<double>(dot(vector,vector)));
+template <typename T, uint64_t N>
+double norm(const StaticVector<T, N> &vector) {
+  return sqrt(static_cast<double>(dot(vector, vector)));
 }
-}
+} // namespace math
